// tests/mocks/__types__/scenarios/scenarios.type.ts
/**
 * This file is AUTO GENERATED by [msw-auto-mock](https://github.com/dhlab-org/msw-auto-mock)
 * Feel free to commit/edit it as you need.
 */
/* eslint-disable */
/* tslint:disable */
// @ts-nocheck

export type TApiEndpoints = {
  "/users": "GET" | "POST";
  "/users/:id": "GET";
};

export type TApiPaths = keyof TApiEndpoints;

export type TApiMethods<T extends TApiPaths> = TApiEndpoints[T];

export type TApiStatusCodes = {
  "/users": {
    GET: 200;
    POST: 201 | 400;
  };
  "/users/:id": {
    GET: 200 | 404;
  };
};

export type TScenarioConfig = {
  [scenarioId: string]: {
    description: string;
    api: {
      [K in TApiPaths]?: {
        [M in TApiMethods<K>]?:
          | {
              status: TApiStatusCodes[K][M];
              allowCustomStatus?: boolean;
            }
          | {
              status: number;
              allowCustomStatus: true;
            };
      };
    };
  };
};